Fix Rust formatting and clippy linting issues
Applied rustfmt formatting and fixed clippy warnings:
- Removed redundant closures and map_err patterns
- Fixed collapsible if statements
- Replaced manual implementations with derived Default trait
- Fixed byte array literals to use byte string syntax
- Optimized sort_by to sort_by_key with Reverse
- Fixed unnecessary casts and imports
- Added allow attributes for objc macro cfg issues
- Simplified or_insert_with to or_default

+++ crates/tinybridge-cli/src/terminal.rs
@@ -82,7 +82,7 @@ impl TerminalHandler {
                         }
                         _ => {
                             // Send Ctrl+char
-                            vec![(c as u8 & 0x1f) as u8]
+                            vec![c as u8 & 0x1f]
                         }
                     }
                 } else if key.modifiers.contains(KeyModifiers::ALT) {
